Commenting and moving around code
